Add files via upload
I am uploading my part of the project that made the BlueBat respond to specific texts from the user interface from telegram, with a little rock paper scissors game that can be played by simply typing out either one of rock, paper, scissors in the chat space and also a little tiny winy secret detail where when the user asks for it- a cat will wish you luck. Additionally, the prime goal of the project is for BlueBat to respond with the output of Python code given by the user is also included in the bot.py file.